requirements.txt
transformers
rank_bm25
wikipedia-api